# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Blockchain Metadata Chain of Custody & Evidence Verification API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Evidence {
  id: number;
  caseId: string;
  title: string;
  /** @nullable */
  description?: string | null;
  fileHash: string;
  /** @nullable */
  blockchainTxHash?: string | null;
  /** @nullable */
  blockNumber?: number | null;
  status: string;
  evidenceType: string;
  collectedBy: string;
  collectedAt: string;
  /** @nullable */
  locationData?: string | null;
  /** @nullable */
  fileSize?: number | null;
  /** @nullable */
  mimeType?: string | null;
  isVerified: boolean;
  accessCount: number;
  createdAt: string;
  updatedAt: string;
}

export interface CreateEvidenceBody {
  caseId: string;
  title: string;
  /** @nullable */
  description?: string | null;
  fileHash: string;
  evidenceType: string;
  collectedBy: string;
  collectedAt: string;
  /** @nullable */
  locationData?: string | null;
  /** @nullable */
  fileSize?: number | null;
  /** @nullable */
  mimeType?: string | null;
}

export interface UpdateEvidenceBody {
  title?: string;
  /** @nullable */
  description?: string | null;
  status?: string;
  /** @nullable */
  locationData?: string | null;
}

export interface CustodyTransaction {
  id: number;
  evidenceId: number;
  /** @nullable */
  evidenceTitle?: string | null;
  /** @nullable */
  caseId?: string | null;
  action: string;
  performedBy: string;
  performedAt: string;
  /** @nullable */
  notes?: string | null;
  txHash: string;
  /** @nullable */
  blockNumber?: number | null;
  /** @nullable */
  previousHash?: string | null;
  /** @nullable */
  ipAddress?: string | null;
  createdAt: string;
}

export interface CreateCustodyTransactionBody {
  action: string;
  performedBy: string;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  ipAddress?: string | null;
}

export interface VerificationResult {
  /** @nullable */
  evidenceId?: number | null;
  hash: string;
  isValid: boolean;
  blockchainConfirmed: boolean;
  /** @nullable */
  blockNumber?: number | null;
  timestamp: string;
  chainIntegrity: boolean;
  totalTransactions: number;
  message: string;
}

export interface LegalCertificate {
  certificateId: string;
  evidenceId: number;
  caseId: string;
  title: string;
  fileHash: string;
  /** @nullable */
  blockchainTxHash?: string | null;
  collectedBy: string;
  collectedAt: string;
  isVerified: boolean;
  chainIntegrity: boolean;
  totalCustodyEvents: number;
  generatedAt: string;
  custodyChain: CustodyTransaction[];
}

export interface CountByLabel {
  label: string;
  count: number;
}

export interface DashboardStats {
  totalEvidence: number;
  verifiedEvidence: number;
  pendingEvidence: number;
  totalCases: number;
  totalTransactions: number;
  integrityRate: number;
  evidenceByType: CountByLabel[];
  evidenceByStatus: CountByLabel[];
}

export interface VerifyHashBody {
  hash: string;
}

export type ListEvidenceParams = {
  status?: string;
  caseId?: string;
};