# lib/api-zod/src/generated/types/countByLabel.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Blockchain Metadata Chain of Custody & Evidence Verification API
 * OpenAPI spec version: 0.1.0
 */

export interface CountByLabel {
  label: string;
  count: number;
}