# lib/api-zod/src/generated/types/createEvidenceBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Blockchain Metadata Chain of Custody & Evidence Verification API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateEvidenceBody {
  caseId: string;
  title: string;
  /** @nullable */
  description?: string | null;
  fileHash: string;
  evidenceType: string;
  collectedBy: string;
  collectedAt: string;
  /** @nullable */
  locationData?: string | null;
  /** @nullable */
  fileSize?: number | null;
  /** @nullable */
  mimeType?: string | null;
}