# lib/api-zod/src/generated/types/custodyTransaction.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Blockchain Metadata Chain of Custody & Evidence Verification API
 * OpenAPI spec version: 0.1.0
 */

export interface CustodyTransaction {
  id: number;
  evidenceId: number;
  /** @nullable */
  evidenceTitle?: string | null;
  /** @nullable */
  caseId?: string | null;
  action: string;
  performedBy: string;
  performedAt: string;
  /** @nullable */
  notes?: string | null;
  txHash: string;
  /** @nullable */
  blockNumber?: number | null;
  /** @nullable */
  previousHash?: string | null;
  /** @nullable */
  ipAddress?: string | null;
  createdAt: string;
}