# lib/api-zod/src/generated/types/dashboardStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Blockchain Metadata Chain of Custody & Evidence Verification API
 * OpenAPI spec version: 0.1.0
 */
import type { CountByLabel } from "./countByLabel";

export interface DashboardStats {
  totalEvidence: number;
  verifiedEvidence: number;
  pendingEvidence: number;
  totalCases: number;
  totalTransactions: number;
  integrityRate: number;
  evidenceByType: CountByLabel[];
  evidenceByStatus: CountByLabel[];
}