# lib/api-zod/src/generated/types/legalCertificate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Blockchain Metadata Chain of Custody & Evidence Verification API
 * OpenAPI spec version: 0.1.0
 */
import type { CustodyTransaction } from "./custodyTransaction";

export interface LegalCertificate {
  certificateId: string;
  evidenceId: number;
  caseId: string;
  title: string;
  fileHash: string;
  /** @nullable */
  blockchainTxHash?: string | null;
  collectedBy: string;
  collectedAt: string;
  isVerified: boolean;
  chainIntegrity: boolean;
  totalCustodyEvents: number;
  generatedAt: string;
  custodyChain: CustodyTransaction[];
}