# lib/api-zod/src/generated/types/updateEvidenceBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Blockchain Metadata Chain of Custody & Evidence Verification API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateEvidenceBody {
  title?: string;
  /** @nullable */
  description?: string | null;
  status?: string;
  /** @nullable */
  locationData?: string | null;
}