# lib/api-zod/src/generated/types/verificationResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Blockchain Metadata Chain of Custody & Evidence Verification API
 * OpenAPI spec version: 0.1.0
 */

export interface VerificationResult {
  /** @nullable */
  evidenceId?: number | null;
  hash: string;
  isValid: boolean;
  blockchainConfirmed: boolean;
  /** @nullable */
  blockNumber?: number | null;
  timestamp: string;
  chainIntegrity: boolean;
  totalTransactions: number;
  message: string;
}